Fix even more tests
