+++ lib/mockzilla-management/src/commonTest/kotlin/testutils/ConnectionConfigImpl.kt
@@ -1,0 +1,5 @@
+package com.apadmi.mockzilla.testutils
+
+import com.apadmi.mockzilla.management.MockzillaManagement
+
+data class ConnectionConfigImpl(override val ip: String, override val port: String) : MockzillaManagement.ConnectionConfig
